+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 # Changelog
-All significant changes will be аfixed in this file.
+All significant changes will be fixed in this file.
+
+
+[0.1.0]
+Frame order.py, status.py, products.py
+Frame errors and tests.